Refactor CSS to styled-components and implement theme switch
The commit refactors the CSS into styled-components in the Header, Main and Card components. It also implements a light and dark theme switch in the Header component. These changes enable a more perceptive user experience and a cleaner, lower-maintenance codebase. React Strict Mode was applied for optimal rendering and better developer warnings.